third_party/blink/web_tests/wpt_internal/prerender/activate-from-iframe.html

<!DOCTYPE html>
<!--
This file cannot be upstreamed to WPT until:
* startPrerendering() usage is replaced with a WebDriver API
-->
<title>Activate a prerendered page from iframe via window.parent.location</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>

promise_test(async t => {
  const resultChannel = new BroadcastChannel('result-channel');
  t.add_cleanup(_ => resultChannel.close());

  const url = `resources/activate-from-iframe.html?mode=triggering`;
  window.open(url, '_blank', 'noopener');

  const gotActivationMessage =
      await new Promise(r => resultChannel.onmessage = r);
  assert_equals(gotActivationMessage.data,
      'attempting prerender activation');
  const gotReloadMessage =
      await new Promise(r => resultChannel.onmessage = r);
  assert_equals(gotReloadMessage.data, 'loaded from scratch');
}, `Navigation on the main frame initiated by window opened by an ` +
    `iframe opening via window.parent.location should not activate a ` +
    `prerendered page`);

</script>
