third_party/blink/web_tests/platform/mac-mac12-arm64/virtual/prerender/wpt_internal/prerender/activate-from-iframe-expected.txt

This is a testharness.js-based test.
FAIL Navigation on the main frame initiated by window opened by an iframe opening via window.parent.location should not activate a prerendered page assert_equals: expected "attempting prerender activation" but got "loaded from scratch"
Harness: the test ran to completion.
